figured runaway frequency issue at the expense

+++ ParseConfig.py
@@ -29,7 +29,7 @@ def load_nodes_from_config(path):
                     
                 nodes[nj["id"]] = (Node(nj["id"],
                             PIDController(nj["id"], float(ctrl_opts["kp"]), float(ctrl_opts["ki"]), int(ctrl_opts["ki_window"]), float(ctrl_opts["kd"]), 
-                                          int(ctrl_opts["diff_step"]), int(ctrl_opts["midpoint"]), int(ctrl_opts["offset"])), all_buffs, float(nj["frequency"])))
+                                          int(ctrl_opts["diff_step"]), int(ctrl_opts["midpoint"]), float(ctrl_opts["offset"])), all_buffs, float(nj["frequency"])))
             else:
                 print("Unknown control scheme " + str(ctrl_opts["type"]))
                 exit(0)
